training/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0
pyyaml>=6.0
azure-storage-blob>=12.0
azure-identity>=1.14
